requirements.txt
Django==2.1.5
djangorestframework==3.9.1
gunicorn==19.9.0
psycopg2==2.7.7

